create Update class
